+++ src/Paket.Core/FindReferences.fs
@@ -4,13 +4,14 @@ open System
 open System.IO
 open Logging
 open Paket.Domain
+open Paket.Rop
 
 let FindReferencesForPackage (dependenciesFileName, package:PackageName) =
     let root = Path.GetDirectoryName dependenciesFileName
     let projectFiles = ProjectFile.FindAllProjects root
     let lockFile = LockFile.LoadFrom((DependenciesFile.FindLockfile dependenciesFileName).FullName)
 
-    [for project,referencesFile in InstallProcess.findAllReferencesFiles root do
+    [for project,referencesFile in InstallProcess.findAllReferencesFiles root |> returnOrFail do
         let installedPackages = 
             referencesFile
             |> lockFile.GetPackageHull